+++ conda_pip/main.py
@@ -1,9 +1,7 @@
 import os
-import sys
-import sysconfig
 from logging import getLogger
 from pathlib import Path
-from subprocess import run, check_output
+from subprocess import run
 from typing import Iterable
 
 try:
@@ -12,51 +10,16 @@ except ImportError:
     from importlib_resources import files as importlib_files
 
 from conda.history import History
-from conda.base.context import context, locate_prefix_by_name
+from conda.base.context import context
 from conda.core.prefix_data import PrefixData
 from conda.cli.python_api import run_command
 from conda.exceptions import CondaError, CondaSystemExit
 from conda.models.match_spec import MatchSpec
 
+from .utils import get_env_python, get_externally_managed_path
+
 logger = getLogger(f"conda.{__name__}")
 HERE = Path(__file__).parent.resolve()
-
-
-def get_prefix(prefix: os.PathLike = None, name: str = None) -> Path:
-    if prefix:
-        return Path(prefix)
-    elif name:
-        return Path(locate_prefix_by_name(name))
-    else:
-        return Path(context.target_prefix)
-
-
-def get_env_python(prefix: os.PathLike = None) -> Path:
-    prefix = Path(prefix or sys.prefix)
-    if os.name == "nt":
-        return prefix / "python.exe"
-    return prefix / "bin" / "python"
-
-
-def get_env_stdlib(prefix: os.PathLike = None) -> Path:
-    prefix = Path(prefix or sys.prefix)
-    if str(prefix) == sys.prefix:
-        return Path(sysconfig.get_path("stdlib"))
-    return Path(check_output([get_env_python(prefix), "-c", "import sysconfig; print(sysconfig.get_paths()['stdlib'])"], text=True).strip())
-
-
-def get_externally_managed_path(prefix: os.PathLike = None) -> Path:
-    prefix = Path(prefix or sys.prefix)
-    if os.name == "nt":
-        yield Path(prefix, "Lib", "EXTERNALLY-MANAGED")
-    else:
-        found = False
-        for python_dir in sorted(Path(prefix, "lib").glob("python*")):
-            if python_dir.is_dir():
-                found = True
-                yield Path(python_dir, "EXTERNALLY-MANAGED")
-        if not found:
-            raise ValueError("Could not locate EXTERNALLY-MANAGED file")
 
 
 def validate_target_env(path: Path, packages: Iterable[str]) -> Iterable[str]:
@@ -67,7 +30,7 @@ def validate_target_env(path: Path, packages: Iterable[str]) -> Iterable[str]:
         raise CondaError(f"Target environment at {path} requires python>=3.2")
     if not list(pd.query("pip>=23.0.1")):
         raise CondaError(f"Target environment at {path} requires pip>=23.0.1")
-
+    
     packages_to_process = []
     for pkg in packages:
         spec = MatchSpec(pkg)
